split apart testminer code
